src/app-server/backend-fastapi/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
passlib
python-jose[cryptography]
python-multipart